Added helpers to reduce code copied over when created scene new scene handlers
