+++ src/app/clusters/scenes-server/SceneTableImpl.cpp
@@ -22,81 +22,6 @@
 namespace chip {
 namespace scenes {
 
-CHIP_ERROR
-DefaultSceneHandlerImpl::EncodeAttributeValueList(const List<AttributeValuePairType> & aVlist, MutableByteSpan & serializedBytes)
-{
-    TLV::TLVWriter writer;
-    writer.Init(serializedBytes);
-    ReturnErrorOnFailure(app::DataModel::Encode(writer, TLV::AnonymousTag(), aVlist));
-    serializedBytes.reduce_size(writer.GetLengthWritten());
-
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR DefaultSceneHandlerImpl::DecodeAttributeValueList(const ByteSpan & serializedBytes,
-                                                             DecodableList<AttributeValuePairDecodableType> & aVlist)
-{
-    TLV::TLVReader reader;
-
-    reader.Init(serializedBytes);
-    ReturnErrorOnFailure(reader.Next(TLV::kTLVType_Array, TLV::AnonymousTag()));
-    ReturnErrorOnFailure(aVlist.Decode(reader));
-
-    return CHIP_NO_ERROR;
-}
-
-CHIP_ERROR
-DefaultSceneHandlerImpl::SerializeAdd(EndpointId endpoint, const ExtensionFieldSetDecodableType & extensionFieldSet,
-                                      MutableByteSpan & serializedBytes)
-{
-    AttributeValuePairType aVPairs[kMaxAvPair];
-
-    size_t pairTotal = 0;
-    // Verify size of list
-    ReturnErrorOnFailure(extensionFieldSet.attributeValueList.ComputeSize(&pairTotal));
-    VerifyOrReturnError(pairTotal <= ArraySize(aVPairs), CHIP_ERROR_BUFFER_TOO_SMALL);
-
-    uint8_t pairCount  = 0;
-    auto pair_iterator = extensionFieldSet.attributeValueList.begin();
-    while (pair_iterator.Next())
-    {
-        aVPairs[pairCount] = pair_iterator.GetValue();
-        pairCount++;
-    }
-    ReturnErrorOnFailure(pair_iterator.GetStatus());
-    List<AttributeValuePairType> attributeValueList(aVPairs, pairCount);
-
-    return EncodeAttributeValueList(attributeValueList, serializedBytes);
-}
-
-CHIP_ERROR DefaultSceneHandlerImpl::Deserialize(EndpointId endpoint, ClusterId cluster, const ByteSpan & serializedBytes,
-                                                ExtensionFieldSetType & extensionFieldSet)
-{
-    DecodableList<AttributeValuePairDecodableType> attributeValueList;
-
-    ReturnErrorOnFailure(DecodeAttributeValueList(serializedBytes, attributeValueList));
-
-    // Verify size of list
-    size_t pairTotal = 0;
-    ReturnErrorOnFailure(attributeValueList.ComputeSize(&pairTotal));
-    VerifyOrReturnError(pairTotal <= ArraySize(mAVPairs), CHIP_ERROR_BUFFER_TOO_SMALL);
-
-    uint8_t pairCount  = 0;
-    auto pair_iterator = attributeValueList.begin();
-    while (pair_iterator.Next())
-    {
-        mAVPairs[pairCount] = pair_iterator.GetValue();
-        pairCount++;
-    };
-    ReturnErrorOnFailure(pair_iterator.GetStatus());
-
-    extensionFieldSet.clusterID          = cluster;
-    extensionFieldSet.attributeValueList = mAVPairs;
-    extensionFieldSet.attributeValueList.reduce_size(pairCount);
-
-    return CHIP_NO_ERROR;
-}
-
 /// @brief Tags Used to serialize Scenes so they can be stored in flash memory.
 /// kSceneCount: Number of scenes in a Fabric
 /// kStorageIDArray: Array of StorageID struct
@@ -817,12 +742,12 @@ CHIP_ERROR DefaultSceneTableImpl::SceneSaveEFS(SceneTableEntry & scene)
         // TODO : Once zap supports the scenable quality, implement a GetSceneableClusterCountFromEndpointType function to avoid
         // over-allocation
         uint8_t clusterCount = GetClusterCountFromEndpoint();
-        chip::Platform::ScopedMemoryBuffer<clusterId> cBuffer;
+        chip::Platform::ScopedMemoryBuffer<ClusterId> cBuffer;
         VerifyOrReturnError(cBuffer.Calloc(clusterCount), CHIP_ERROR_NO_MEMORY);
         clusterCount = GetClustersFromEndpoint(cBuffer.Get(), clusterCount);
 
-        Span<clusterId> cSpan(cBuffer.Get(), clusterCount);
-        for (clusterId cluster : cSpan)
+        Span<ClusterId> cSpan(cBuffer.Get(), clusterCount);
+        for (ClusterId cluster : cSpan)
         {
             ExtensionFieldSet EFS;
             MutableByteSpan EFSSpan = MutableByteSpan(EFS.mBytesBuffer, kMaxFieldBytesPerCluster);
