+++ src/app/clusters/scenes-server/SceneTableImpl.h
@@ -17,6 +17,7 @@
 
 #pragma once
 #include <app/clusters/scenes-server/ExtensionFieldSetsImpl.h>
+#include <app/clusters/scenes-server/SceneHandlerImpl.h>
 #include <app/clusters/scenes-server/SceneTable.h>
 #include <app/util/attribute-storage.h>
 #include <lib/core/DataModelTypes.h>
@@ -39,69 +40,6 @@ static_assert(kMaxScenesPerEndpoint <= CHIP_CONFIG_MAX_SCENES_TABLE_SIZE,
               "CHIP_CONFIG_MAX_SCENES_TABLE_SIZE in CHIPConfig.h if you really need more scenes");
 static_assert(kMaxScenesPerEndpoint >= 16, "Per spec, kMaxScenesPerEndpoint must be at least 16");
 static constexpr uint16_t kMaxScenesPerFabric = (kMaxScenesPerEndpoint - 1) / 2;
-
-using clusterId = chip::ClusterId;
-
-/// @brief Default implementation of handler, handle EFS from add scene and view scene commands for any cluster
-///        The implementation of SerializeSave and ApplyScene were omitted and must be implemented in a way that
-///        is compatible with the SerializeAdd output in order to function with the Default Scene Handler.
-///        It is worth noting that this implementation is very memory consuming. In the current worst case,
-///        (Color control cluster), the Extension Field Set's value pair list TLV occupies 99 bytes of memory
-class DefaultSceneHandlerImpl : public scenes::SceneHandler
-{
-
-    template <typename T>
-    using List = chip::app::DataModel::List<T>;
-
-    template <typename T>
-    using DecodableList = chip::app::DataModel::DecodableList<T>;
-
-    using AttributeValuePairType          = chip::app::Clusters::Scenes::Structs::AttributeValuePair::Type;
-    using AttributeValuePairDecodableType = chip::app::Clusters::Scenes::Structs::AttributeValuePair::DecodableType;
-    using ExtensionFieldSetDecodableType  = chip::app::Clusters::Scenes::Structs::ExtensionFieldSet::DecodableType;
-    using ExtensionFieldSetType           = chip::app::Clusters::Scenes::Structs::ExtensionFieldSet::Type;
-
-public:
-    static constexpr uint8_t kMaxAvPair = CHIP_CONFIG_SCENES_MAX_AV_PAIRS_EFS;
-
-    DefaultSceneHandlerImpl() = default;
-    ~DefaultSceneHandlerImpl() override{};
-
-    /// @brief Encodes an attribute value list into a TLV structure and resizes the buffer to the size of the encoded data
-    /// @param aVlist[in] Attribute value list to encode
-    /// @param serializedBytes[out] Buffer to fill from the Attribute value list in a TLV format
-    /// @return CHIP_ERROR
-    virtual CHIP_ERROR EncodeAttributeValueList(const List<AttributeValuePairType> & aVlist, MutableByteSpan & serializedBytes);
-
-    /// @brief Decodes an attribute value list from a TLV structure and ensure it fits the member pair buffer
-    /// @param serializedBytes [in] Buffer to read from
-    /// @param aVlist [out] Attribute value list to fill from the TLV structure.  Only valid while the buffer backing
-    /// serializedBytes exists and its contents are not modified.
-    /// @return CHIP_ERROR
-    virtual CHIP_ERROR DecodeAttributeValueList(const ByteSpan & serializedBytes,
-                                                DecodableList<AttributeValuePairDecodableType> & aVlist);
-
-    /// @brief From command AddScene, allows handler to filter through clusters in command to serialize only the supported ones.
-    /// @param endpoint[in] Endpoint ID
-    /// @param extensionFieldSet[in] ExtensionFieldSets provided by the AddScene Command, pre initialized
-    /// @param serializedBytes[out] Buffer to fill from the ExtensionFieldSet in command
-    /// @return CHIP_NO_ERROR if successful, CHIP_ERROR_INVALID_ARGUMENT if the cluster is not supported, CHIP_ERROR value
-    /// otherwise
-    virtual CHIP_ERROR SerializeAdd(EndpointId endpoint, const ExtensionFieldSetDecodableType & extensionFieldSet,
-                                    MutableByteSpan & serializedBytes) override;
-
-    /// @brief Simulates taking data from nvm and loading it in a command object if the cluster is supported by the endpoint
-    /// @param endpoint target endpoint
-    /// @param cluster  target cluster
-    /// @param serializedBytes data to deserialize into EFS
-    /// @return CHIP_NO_ERROR if Extension Field Set was successfully populated, CHIP_ERROR_INVALID_ARGUMENT if the cluster is not
-    /// supported, specific CHIP_ERROR otherwise
-    virtual CHIP_ERROR Deserialize(EndpointId endpoint, ClusterId cluster, const ByteSpan & serializedBytes,
-                                   ExtensionFieldSetType & extensionFieldSet) override;
-
-private:
-    AttributeValuePairType mAVPairs[kMaxAvPair];
-};
 
 /**
  * @brief Implementation of a storage in nonvolatile storage of the scene table.
